+++ apps/openmw/mwscript/skyextensions.cpp
@@ -106,7 +106,7 @@ namespace MWScript
                 chances.reserve(10);
                 while (arg0 > 0)
                 {
-                    chances.push_back(std::clamp(runtime[0].mInteger, 0, 100));
+                    chances.push_back(static_cast<uint8_t>(std::clamp(runtime[0].mInteger, 0, 100)));
                     runtime.pop();
                     arg0--;
                 }
